+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 0.0.2
+
+## New
+- Added Set Custom Rewards support
+- Added Clear Rewards support
+- Added a way to subscribe to channel reward events
+- Added a way to subscribe to channel follows
+- Added a way to subscribe to subscribers
+- Added a way to subscribe to hype train
+- Added a way to subscribe to raids
+
+## Fixed
+- Adjusted folders for files
+- Clean up comments
+
 # 0.0.1
 
 ## New